requirements.txt
requests>=2.31.0
openpyxl>=3.1.2
language-tool-python>=2.7.1
python-dotenv>=1.0.0
pyspellchecker>=0.7.0
